Rename "sys" module to "ipc"
