+++ plib/src/lib.rs
@@ -10,12 +10,12 @@
 pub mod curuser;
 pub mod group;
 pub mod io;
+pub mod ipc;
 pub mod lzw;
 pub mod modestr;
 pub mod platform;
 pub mod priority;
 pub mod sccsfile;
-pub mod sys;
 pub mod terminal;
 pub mod testing;
 pub mod utmpx;
